fix(agents): install pinned agent binaries with the windows executable extension
install_binary_or_archive_from_pin wrote every pinned artifact to
managed_dir.join(kind.as_str()), an extension-less path such as `claude` or
`codex`. Windows will not execute a file without a `.exe` suffix, so with the
Windows pins from the base branch in place the installer would download the
right artifact, verify its sha256, write it somewhere Windows cannot run, and
leave the failure to surface later as a misleading launch error.

Both source arms were affected, for different reasons. The `Binary` arm never
derives a name from the download at all, so claude's `.../win32-x64/claude.exe`
landed as `claude`. The `Archive` arm extracts an inner member whose name
already carries the extension on Windows (codex ships
`codex-x86_64-pc-windows-msvc.exe`) and renames it onto the same target path,
dropping the suffix on the way.

Routes the write site and all three read sites through one
managed_pinned_binary_path helper in readiness::paths, built on the existing
platform_binary_filename convention, which moves from installer::npm to the
shared agent_cli::executable helpers so readiness can reach it. Renaming only
the write site would have converted a launch failure into a not-found failure,
so the idempotent-skip probe, native artifact resolution and the launcher PATH
prefix probe all move with it.

Refs #2141
