+++ anyharness/crates/anyharness-lib/src/integrations/agent_cli/executable.rs
@@ -1,5 +1,18 @@
 use std::path::{Path, PathBuf};
 
+/// The on-disk filename an executable must carry on this platform. Windows
+/// will not execute an extension-less file, so every managed artifact we place
+/// has to be named `<name>.exe` there. Single source of truth for that rule:
+/// the installer writes through it and every resolver reads through it, so a
+/// write site and a read site can never disagree about the name.
+pub(crate) fn platform_binary_filename(binary_name: &str) -> PathBuf {
+    if cfg!(windows) && !binary_name.to_ascii_lowercase().ends_with(".exe") {
+        PathBuf::from(format!("{binary_name}.exe"))
+    } else {
+        PathBuf::from(binary_name)
+    }
+}
+
 pub(crate) fn find_in_path(binary_name: &str) -> Option<PathBuf> {
     find_in_path_matching(binary_name, |_| true)
 }
@@ -67,6 +80,46 @@ pub(crate) fn make_executable(_path: &Path) -> std::io::Result<()> {
 #[cfg(test)]
 mod tests {
     use super::*;
+
+    #[test]
+    fn platform_binary_filename_carries_the_windows_executable_extension() {
+        // Runs on every platform on purpose: the interesting half of this rule
+        // is the Windows half, and a `cfg(windows)`-only test would never run
+        // in a unix-only CI matrix. Asserting the contract in both directions
+        // keeps the unix arm honest too, so nobody "fixes" Windows by
+        // appending `.exe` everywhere.
+        let name = platform_binary_filename("claude");
+
+        assert_eq!(
+            name.file_stem().and_then(|stem| stem.to_str()),
+            Some("claude"),
+            "the stem must always be the requested binary name"
+        );
+        assert_eq!(
+            name.extension().and_then(|ext| ext.to_str()),
+            if cfg!(windows) { Some("exe") } else { None },
+            "windows needs a .exe suffix to execute the file at all; unix must not get one"
+        );
+        assert_eq!(
+            name,
+            PathBuf::from(if cfg!(windows) {
+                "claude.exe"
+            } else {
+                "claude"
+            })
+        );
+    }
+
+    #[test]
+    fn platform_binary_filename_does_not_double_up_an_existing_extension() {
+        // Archive pins name their inner member with the extension already
+        // present (codex ships `codex-x86_64-pc-windows-msvc.exe`). The
+        // destination name is derived from the agent kind, never from that
+        // member, so this helper is only ever handed a bare kind — but if a
+        // caller ever passes a suffixed name, we must not produce `x.exe.exe`.
+        let name = platform_binary_filename("codex.exe");
+        assert_eq!(name, PathBuf::from("codex.exe"));
+    }
 
     #[test]
     fn detects_superset_agent_wrapper_marker() {
